Add files via upload
Project in C where I create a basic chat program using the TCP protocol and multi-threading programming. It isn't secure or industry standard at all, but it works between clients from different networks, with the help the ssh command in linux. More in the first README file. (Instructions are in Romanian for now).